Prod koyeb

+++ src/projects/projects.service.ts
@@ -1,14 +1,24 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Project } from './project.entity';
 import { User } from '../users/user.entity'; // Import User entity
+import { Request } from 'src/requests/requests.entity';
+import { UpdateProjectDto } from './dto/update-project.dto';
+import { CreateProjectDto } from './dto/create-project.dto';
 
 @Injectable()
 export class ProjectsService {
   constructor(
     @InjectRepository(Project)
     private projectsRepository: Repository<Project>,
+
+    @InjectRepository(Request)
+    private requestsRepository: Repository<Request>,
 
     @InjectRepository(User)
     private usersRepository: Repository<User>, // Inject User repository not best solution ?
@@ -26,17 +36,25 @@ export class ProjectsService {
   }
 
   // Create a new project with assigned foreman
-  async create(
-    projectData: Partial<Project>,
-    foremanId?: number,
-  ): Promise<Project> {
-    const newProject = this.projectsRepository.create(projectData);
+  async create(createProjectDto: CreateProjectDto): Promise<Project> {
+    const {
+      name,
+      description,
+      status,
+      assignedForeman = null,
+    } = createProjectDto;
 
-    if (foremanId) {
-      const foreman = await this.usersRepository.findOneBy({ id: foremanId });
-      if (foreman) {
-        newProject.assignedForeman = foreman; // Assign the User entity as foreman
-      }
+    const newProject = this.projectsRepository.create({
+      name,
+      description,
+      status,
+    });
+
+    if (assignedForeman) {
+      const foreman = await this.usersRepository.findOneBy({
+        id: assignedForeman,
+      });
+      newProject.assignedForeman = foreman || null; // Assign foreman if found, or set to null
     }
 
     return this.projectsRepository.save(newProject);
@@ -45,8 +63,7 @@ export class ProjectsService {
   // Update a project and handle assigned foreman
   async update(
     id: number,
-    projectData: Partial<Project>,
-    foremanId?: number,
+    updateProjectDto: UpdateProjectDto,
   ): Promise<Project> {
     const project = await this.projectsRepository.findOne({
       where: { id },
@@ -56,22 +73,46 @@ export class ProjectsService {
     if (!project) {
       throw new Error(`Project with ID ${id} not found.`);
     }
+    const foremanId = updateProjectDto.assignedForeman;
 
-    Object.assign(project, projectData); // Update basic project data
-
-    if (foremanId) {
+    if (foremanId && foremanId != -1) {
       const foreman = await this.usersRepository.findOneBy({ id: foremanId });
       if (foreman) {
         project.assignedForeman = foreman; // Update the foreman
       } else {
         throw new Error(`Foreman with ID ${foremanId} not found.`);
       }
+    } else {
+      project.assignedForeman = null;
     }
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const { assignedForeman, ...otherUpdates } = updateProjectDto;
+    Object.assign(project, otherUpdates);
 
     return this.projectsRepository.save(project); // Save updated project with foreman
   }
 
   async delete(id: number): Promise<void> {
+    const project = await this.projectsRepository.findOne({
+      where: { id },
+    });
+
+    if (!project) {
+      throw new NotFoundException('Project not found');
+    }
+
+    // Use QueryBuilder to count requests associated with the project
+    const requestCount = await this.requestsRepository
+      .createQueryBuilder('request')
+      .where('request.projectId = :projectId', { projectId: id })
+      .getCount();
+
+    if (requestCount > 0) {
+      throw new BadRequestException(
+        'Cannot delete project with associated requests. Please delete or reassign requests first.',
+      );
+    }
+
     await this.projectsRepository.delete(id);
   }
 }
